% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{use_sentence_linebreaks}
\alias{use_sentence_linebreaks}
\title{Set Visual editor's Markdown wrapping option to "Sentence" (RStudio IDE).}
\usage{
use_sentence_linebreaks(scope = c("user", "project"))
}
\arguments{
\item{scope}{Edit globally for the current \emph{user},
or locally for the current \emph{project}.}
}
\value{
Nothing, used for its side-effects.
}
\description{
Set Visual editor's Markdown wrapping option to "Sentence" (RStudio IDE).
}
